+++ src/backend/mysql/table.rs
@@ -134,9 +134,12 @@ impl TableBuilder for MysqlQueryBuilder {
                 write!(sql, ", ").unwrap();
             };
             match option {
-                TableAlterOption::AddColumn(column_def) => {
+                TableAlterOption::AddColumn(AddColumnOption {
+                    column,
+                    if_not_exists: _,
+                }) => {
                     write!(sql, "ADD COLUMN ").unwrap();
-                    self.prepare_column_def(column_def, sql);
+                    self.prepare_column_def(column, sql);
                 }
                 TableAlterOption::ModifyColumn(column_def) => {
                     write!(sql, "MODIFY COLUMN ").unwrap();