+++ src/query/mod.rs
@@ -10,20 +10,24 @@
 mod condition;
 mod delete;
 mod insert;
+mod on_conflict;
 mod ordered;
 mod select;
 mod shim;
 mod traits;
 mod update;
+mod window;
 mod with;
 
 pub use condition::*;
 pub use delete::*;
 pub use insert::*;
+pub use on_conflict::*;
 pub use ordered::*;
 pub use select::*;
 pub use traits::*;
 pub use update::*;
+pub use window::*;
 pub use with::*;
 
 /// Shorthand for constructing any table query