+++ src/query/select.rs
@@ -5,7 +5,8 @@ use crate::{
     query::{condition::*, OrderedStatement},
     types::*,
     value::*,
-    QueryStatementBuilder, QueryStatementWriter, SubQueryStatement, WithClause, WithQuery,
+    QueryStatementBuilder, QueryStatementWriter, SubQueryStatement, WindowStatement, WithClause,
+    WithQuery,
 };
 
 /// Select rows from an existing table
@@ -51,6 +52,7 @@ pub struct SelectStatement {
     pub(crate) limit: Option<Value>,
     pub(crate) offset: Option<Value>,
     pub(crate) lock: Option<LockClause>,
+    pub(crate) window: Option<(DynIden, WindowStatement)>,
 }
 
 /// List of distinct keywords that can be used in select statement
@@ -61,11 +63,21 @@ pub enum SelectDistinct {
     DistinctRow,
 }
 
+/// Window type in [`SelectExpr`]
+#[derive(Debug, Clone)]
+pub enum WindowSelectType {
+    /// Name in [`SelectStatement`]
+    Name(DynIden),
+    /// Inline query in [`SelectExpr`]
+    Query(WindowStatement),
+}
+
 /// Select expression used in select statement
 #[derive(Debug, Clone)]
 pub struct SelectExpr {
     pub expr: SimpleExpr,
     pub alias: Option<DynIden>,
+    pub window: Option<WindowSelectType>,
 }
 
 /// Join expression used in select statement
@@ -113,6 +125,7 @@ impl Into<SelectExpr> for SimpleExpr {
         SelectExpr {
             expr: self,
             alias: None,
+            window: None,
         }
     }
 }
@@ -139,6 +152,7 @@ impl SelectStatement {
             limit: None,
             offset: None,
             lock: None,
+            window: None,
         }
     }
 
@@ -157,6 +171,7 @@ impl SelectStatement {
             limit: self.limit.take(),
             offset: self.offset.take(),
             lock: self.lock.take(),
+            window: self.window.take(),
         }
     }
 
@@ -501,6 +516,7 @@ impl SelectStatement {
         self.expr(SelectExpr {
             expr: expr.into(),
             alias: Some(alias.into_iden()),
+            window: None,
         });
         self
     }
@@ -515,6 +531,160 @@ impl SelectStatement {
         A: IntoIden,
     {
         self.expr_as(expr, alias)
+    }
+
+    /// Select column with window function.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use sea_query::{tests_cfg::*, *};
+    ///
+    /// let query = Query::select()
+    ///     .from(Char::Table)
+    ///     .expr_window(Expr::col(Char::Character), WindowStatement::partition_by(Char::FontSize))
+    ///     .to_owned();
+    ///
+    /// assert_eq!(
+    ///     query.to_string(MysqlQueryBuilder),
+    ///     r#"SELECT `character` OVER ( PARTITION BY `font_size` ) FROM `character`"#
+    /// );
+    /// assert_eq!(
+    ///     query.to_string(PostgresQueryBuilder),
+    ///     r#"SELECT "character" OVER ( PARTITION BY "font_size" ) FROM "character""#
+    /// );
+    /// assert_eq!(
+    ///     query.to_string(SqliteQueryBuilder),
+    ///     r#"SELECT "character" OVER ( PARTITION BY "font_size" ) FROM "character""#
+    /// );
+    /// ```
+    pub fn expr_window<T>(&mut self, expr: T, window: WindowStatement) -> &mut Self
+    where
+        T: Into<SimpleExpr>,
+    {
+        self.expr(SelectExpr {
+            expr: expr.into(),
+            alias: None,
+            window: Some(WindowSelectType::Query(window)),
+        });
+        self
+    }
+
+    /// Select column with window function and label.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use sea_query::{tests_cfg::*, *};
+    ///
+    /// let query = Query::select()
+    ///     .from(Char::Table)
+    ///     .expr_window_as(Expr::col(Char::Character), WindowStatement::partition_by(Char::FontSize), Alias::new("C"))
+    ///     .to_owned();
+    ///
+    /// assert_eq!(
+    ///     query.to_string(MysqlQueryBuilder),
+    ///     r#"SELECT `character` OVER ( PARTITION BY `font_size` ) AS `C` FROM `character`"#
+    /// );
+    /// assert_eq!(
+    ///     query.to_string(PostgresQueryBuilder),
+    ///     r#"SELECT "character" OVER ( PARTITION BY "font_size" ) AS "C" FROM "character""#
+    /// );
+    /// assert_eq!(
+    ///     query.to_string(SqliteQueryBuilder),
+    ///     r#"SELECT "character" OVER ( PARTITION BY "font_size" ) AS "C" FROM "character""#
+    /// );
+    /// ```
+    pub fn expr_window_as<T, A>(&mut self, expr: T, window: WindowStatement, alias: A) -> &mut Self
+    where
+        T: Into<SimpleExpr>,
+        A: IntoIden,
+    {
+        self.expr(SelectExpr {
+            expr: expr.into(),
+            alias: Some(alias.into_iden()),
+            window: Some(WindowSelectType::Query(window)),
+        });
+        self
+    }
+
+    /// Select column with window name.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use sea_query::{tests_cfg::*, *};
+    ///
+    /// let query = Query::select()
+    ///     .from(Char::Table)
+    ///     .expr_window_name(Expr::col(Char::Character), Alias::new("w"))
+    ///     .window(Alias::new("w"), WindowStatement::partition_by(Char::FontSize))
+    ///     .to_owned();
+    ///
+    /// assert_eq!(
+    ///     query.to_string(MysqlQueryBuilder),
+    ///     r#"SELECT `character` OVER `w` FROM `character` WINDOW `w` AS PARTITION BY `font_size`"#
+    /// );
+    /// assert_eq!(
+    ///     query.to_string(PostgresQueryBuilder),
+    ///     r#"SELECT "character" OVER "w" FROM "character" WINDOW "w" AS PARTITION BY "font_size""#
+    /// );
+    /// assert_eq!(
+    ///     query.to_string(SqliteQueryBuilder),
+    ///     r#"SELECT "character" OVER "w" FROM "character" WINDOW "w" AS PARTITION BY "font_size""#
+    /// );
+    /// ```
+    pub fn expr_window_name<T, W>(&mut self, expr: T, window: W) -> &mut Self
+    where
+        T: Into<SimpleExpr>,
+        W: IntoIden,
+    {
+        self.expr(SelectExpr {
+            expr: expr.into(),
+            alias: None,
+            window: Some(WindowSelectType::Name(window.into_iden())),
+        });
+        self
+    }
+
+    /// Select column with window name and label.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use sea_query::{tests_cfg::*, *};
+    ///
+    /// let query = Query::select()
+    ///     .from(Char::Table)
+    ///     .expr_window_name_as(Expr::col(Char::Character), Alias::new("w"), Alias::new("C"))
+    ///     .window(Alias::new("w"), WindowStatement::partition_by(Char::FontSize))
+    ///     .to_owned();
+    ///
+    /// assert_eq!(
+    ///     query.to_string(MysqlQueryBuilder),
+    ///     r#"SELECT `character` OVER `w` AS `C` FROM `character` WINDOW `w` AS PARTITION BY `font_size`"#
+    /// );
+    /// assert_eq!(
+    ///     query.to_string(PostgresQueryBuilder),
+    ///     r#"SELECT "character" OVER "w" AS "C" FROM "character" WINDOW "w" AS PARTITION BY "font_size""#
+    /// );
+    /// assert_eq!(
+    ///     query.to_string(SqliteQueryBuilder),
+    ///     r#"SELECT "character" OVER "w" AS "C" FROM "character" WINDOW "w" AS PARTITION BY "font_size""#
+    /// );
+    /// ```
+    pub fn expr_window_name_as<T, W, A>(&mut self, expr: T, window: W, alias: A) -> &mut Self
+    where
+        T: Into<SimpleExpr>,
+        A: IntoIden,
+        W: IntoIden,
+    {
+        self.expr(SelectExpr {
+            expr: expr.into(),
+            alias: Some(alias.into_iden()),
+            window: Some(WindowSelectType::Name(window.into_iden())),
+        });
+        self
     }
 
     /// From table.
@@ -1906,6 +2076,40 @@ impl SelectStatement {
     /// ```
     pub fn with(self, clause: WithClause) -> WithQuery {
         clause.query(self)
+    }
+
+    /// WINDOW
+    ///
+    /// # Examples:
+    ///
+    /// ```
+    /// use sea_query::{tests_cfg::*, *};
+    ///
+    /// let query = Query::select()
+    ///     .from(Char::Table)
+    ///     .expr_window_name_as(Expr::col(Char::Character), Alias::new("w"), Alias::new("C"))
+    ///     .window(Alias::new("w"), WindowStatement::partition_by(Char::FontSize))
+    ///     .to_owned();
+    ///
+    /// assert_eq!(
+    ///     query.to_string(MysqlQueryBuilder),
+    ///     r#"SELECT `character` OVER `w` AS `C` FROM `character` WINDOW `w` AS PARTITION BY `font_size`"#
+    /// );
+    /// assert_eq!(
+    ///     query.to_string(PostgresQueryBuilder),
+    ///     r#"SELECT "character" OVER "w" AS "C" FROM "character" WINDOW "w" AS PARTITION BY "font_size""#
+    /// );
+    /// assert_eq!(
+    ///     query.to_string(SqliteQueryBuilder),
+    ///     r#"SELECT "character" OVER "w" AS "C" FROM "character" WINDOW "w" AS PARTITION BY "font_size""#
+    /// );
+    /// ```
+    pub fn window<A>(&mut self, name: A, window: WindowStatement) -> &mut Self
+    where
+        A: IntoIden,
+    {
+        self.window = Some((name.into_iden(), window));
+        self
     }
 }